LayoutTests/http/wpt/webcodecs/encoder-task-failing-expected.txt

CONSOLE MESSAGE: VideoEncoder encode failed: Encoding task failed

PASS Test that H264 encoding failure is propagated to JavaScript
